requirements.txt
# Core framework
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-RESTful==0.3.10
Flask-Caching==2.1.0
Flask-SocketIO==5.3.6

# Database
SQLAlchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.0

# Async and WebSockets
eventlet==0.33.3
gevent==23.9.1
gevent-websocket==0.10.1
python-socketio==5.10.0
python-engineio==4.8.0

# Task Queue
celery==5.3.4
redis==5.0.1
Flask-Executor==1.0.0

# Handling Excel files
openpyxl==3.1.2
pandas==2.1.1
xlrd==2.0.1
XlsxWriter==3.1.4

# Web Server
gunicorn==21.2.0
Werkzeug==2.3.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
six==1.16.0
pytz==2023.3.post1
importlib-metadata==6.8.0
zipp==3.17.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0
